+++ components/experimental/clicker-react/clicker-functional/.eslintrc.js
@@ -1,0 +1,11 @@
+/*!
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+module.exports = {
+    "extends": [
+        "@microsoft/eslint-config-fluid"
+    ],
+    "rules": {}
+}
