+++ components/experimental/clicker-react/clicker-functional/.npmignore
@@ -1,0 +1,3 @@
+nyc
+*.log
+**/*.tsbuildinfo
