+++ components/experimental/clicker-react/clicker-functional/tests/clickerWithHooks.test.ts
@@ -1,0 +1,24 @@
+/*!
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+import { globals } from "../jest.config";
+
+describe("clickerWithHooks", () => {
+
+    beforeAll(async () => {
+        // Wait for the page to load first before running any tests
+        // so this time isn't attributed to the first test
+        await page.goto(globals.PATH, { waitUntil: "load" });
+    }, 45000);
+
+    beforeEach(async () => {
+        await page.goto(globals.PATH, { waitUntil: "load" });
+        await page.waitFor(() => window["fluidStarted"]);
+    });
+
+    it("There's a button to be clicked", async () => {
+        await expect(page).toClick("button", { text: "+" });
+    });
+});
