+++ components/experimental/clicker-react/clicker-functional/tsconfig.esnext.json
@@ -1,0 +1,7 @@
+{
+    "extends": "./tsconfig.json",
+    "compilerOptions": {
+        "outDir": "./lib",
+        "module": "esnext"
+    },
+}
