+++ components/experimental/clicker-react/clicker-functional/webpack.prod.js
@@ -1,0 +1,9 @@
+/*!
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+module.exports = {
+    mode: "production",
+    devtool: "source-map"
+};
