+++ components/experimental/clicker-react/clicker-reducer/.npmignore
@@ -1,0 +1,3 @@
+nyc
+*.log
+**/*.tsbuildinfo
