+++ components/experimental/clicker-react/clicker-reducer/package.json
@@ -1,0 +1,95 @@
+{
+  "name": "@fluid-example/clicker-reducer",
+  "version": "0.21.0",
+  "description": "Minimal Fluid component sample to implement a collaborative counter using Fluid React hooks",
+  "repository": "microsoft/FluidFramework",
+  "license": "MIT",
+  "author": "Microsoft",
+  "main": "dist/index.js",
+  "module": "lib/index.js",
+  "types": "dist/index.d.ts",
+  "scripts": {
+    "build": "concurrently npm:build:compile npm:lint",
+    "build:compile": "concurrently npm:tsc npm:build:esnext",
+    "build:esnext": "tsc --project ./tsconfig.esnext.json",
+    "build:full": "concurrently npm:build npm:webpack",
+    "build:full:compile": "concurrently npm:build:compile npm:webpack",
+    "clean": "rimraf dist lib *.tsbuildinfo *.build.log",
+    "deploy": "npm publish --registry https://packages.wu2.prague.office-int.com",
+    "eslint": "eslint --ext=ts,tsx --format stylish src",
+    "eslint:fix": "eslint --ext=ts,tsx --format stylish src --fix",
+    "lint": "npm run eslint",
+    "lint:fix": "npm run eslint:fix",
+    "prepack": "npm run webpack",
+    "start": "webpack-dev-server --config webpack.config.js --package package.json",
+    "start:docker": "webpack-dev-server --config webpack.config.js --package package.json --env.mode docker",
+    "start:r11s": "webpack-dev-server --config webpack.config.js --package package.json --env.mode r11s",
+    "start:spo": "webpack-dev-server --config webpack.config.js --package package.json --env.mode spo",
+    "start:spo-df": "webpack-dev-server --config webpack.config.js --package package.json --env.mode spo-df",
+    "start:tinylicious": "webpack-dev-server --config webpack.config.js --package package.json --env.mode tinylicious",
+    "test": "npm run test:jest",
+    "test:jest": "jest",
+    "tsc": "tsc",
+    "webpack": "webpack --env.production",
+    "webpack:dev": "webpack --env.development"
+  },
+  "dependencies": {
+    "@fluidframework/aqueduct": "^0.21.0",
+    "@fluidframework/component-core-interfaces": "^0.21.0",
+    "@fluidframework/component-runtime-definitions": "^0.21.0",
+    "@fluidframework/counter": "^0.21.0",
+    "@fluidframework/react": "^0.21.0",
+    "@fluidframework/runtime-definitions": "^0.21.0",
+    "@fluidframework/view-interfaces": "^0.21.0",
+    "react": "^16.10.2",
+    "react-dom": "^16.10.2"
+  },
+  "devDependencies": {
+    "@fluidframework/build-common": "^0.18.0-0",
+    "@fluidframework/webpack-component-loader": "^0.21.0",
+    "@microsoft/eslint-config-fluid": "^0.16.0",
+    "@types/expect-puppeteer": "2.2.1",
+    "@types/jest": "22.2.3",
+    "@types/jest-environment-puppeteer": "2.2.0",
+    "@types/node": "^10.17.24",
+    "@types/puppeteer": "1.3.0",
+    "@types/react": "^16.9.15",
+    "@types/react-dom": "^16.9.4",
+    "@typescript-eslint/eslint-plugin": "~2.17.0",
+    "@typescript-eslint/parser": "~2.17.0",
+    "concurrently": "^5.2.0",
+    "eslint": "~6.8.0",
+    "eslint-plugin-eslint-comments": "~3.1.2",
+    "eslint-plugin-import": "2.20.0",
+    "eslint-plugin-no-null": "~1.0.2",
+    "eslint-plugin-optimize-regex": "~1.1.7",
+    "eslint-plugin-prefer-arrow": "~1.1.7",
+    "eslint-plugin-react": "~7.18.0",
+    "eslint-plugin-unicorn": "~15.0.1",
+    "jest": "^24.9.0",
+    "jest-junit": "^10.0.0",
+    "jest-puppeteer": "^4.3.0",
+    "puppeteer": "^1.20.0",
+    "rimraf": "^2.6.2",
+    "ts-jest": "24.1.0",
+    "ts-loader": "^6.1.2",
+    "typescript": "~3.7.4",
+    "webpack": "^4.43.0",
+    "webpack-dev-server": "^3.8.0",
+    "webpack-merge": "^4.1.4"
+  },
+  "fluid": {
+    "browser": {
+      "umd": {
+        "files": [
+          "dist/main.bundle.js"
+        ],
+        "library": "main"
+      }
+    }
+  },
+  "jest-junit": {
+    "outputDirectory": "nyc",
+    "outputName": "jest-junit-report.xml"
+  }
+}
