+++ components/experimental/clicker-react/clicker-reducer/webpack.dev.js
@@ -1,0 +1,9 @@
+/*!
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+module.exports = {
+    mode: "development",
+    devtool: "inline-source-map"
+};
